+++ spec/lib/facilities/nca_facility_spec.rb
@@ -3,8 +3,8 @@
 require 'rails_helper'
 module Facilities
   RSpec.describe NCAFacility do
-    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
-    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
+    before(:each) { BaseFacility.validate_on_load = false }
+    after(:each) { BaseFacility.validate_on_load = true }
 
     it 'should be an NCAFacility object' do
       expect(described_class.new).to be_an(NCAFacility)
