+++ spec/lib/facilities/vba_facility_spec.rb
@@ -9,8 +9,8 @@ module Facilities
     end
   end
   RSpec.describe VBAFacility do
-    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
-    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
+    before(:each) { BaseFacility.validate_on_load = false }
+    after(:each) { BaseFacility.validate_on_load = true }
 
     it 'should be a VBAFacility object' do
       expect(described_class.new).to be_a(VBAFacility)
