+++ spec/lib/facilities/vc_facility_spec.rb
@@ -3,8 +3,8 @@
 require 'rails_helper'
 module Facilities
   RSpec.describe VCFacility do
-    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
-    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
+    before(:each) { BaseFacility.validate_on_load = false }
+    after(:each) { BaseFacility.validate_on_load = true }
 
     it 'should be a VCFacility object' do
       expect(described_class.new).to be_a(VCFacility)
